update nl_tcd_adc_work.c, nl_tcd_interpol.*
Change type from uint32_t to uint16_t which is way enough for the ribbon interpolation. Interim calculation is still 32 bits.
Reduces calabration data size stored in EEPROM which helps speed and time-slice safety (EEPROM access is slow)

+++ projects/realtime-system/nl_lib/nl_lib_com/tcd/nl_tcd_adc_work.c
@@ -59,9 +59,9 @@ static int      rib_updateEeprom = 0;  // flag / step chain variable
 
 typedef struct
 {
-  int32_t threshold;
-  int32_t tableX[33];
-  int32_t tableY[33];
+  int16_t threshold;
+  int16_t tableX[33];
+  int16_t tableY[33];
 } RibbonCalibrationData_T;
 
 // clang-format off
