+++ projects/realtime-system/nl_lib/nl_lib_com/tcd/nl_tcd_interpol.h
@@ -16,12 +16,12 @@
 typedef struct
 {
   int      points;    ///< number of points in the table, for both arrays !
-  int32_t *x_values;  ///< x-values (input) must be sorted / monotonic !!
-  int32_t *y_values;  ///< y-values (output) can be arbitrary
+  int16_t *x_values;  ///< x-values (input) must be sorted / monotonic !!
+  int16_t *y_values;  ///< y-values (output) can be arbitrary
 } LIB_interpol_data_T;
 
 // Table lookup with interpolation
-int32_t LIB_InterpolateValue(LIB_interpol_data_T *table, int32_t x);
+int16_t LIB_InterpolateValue(LIB_interpol_data_T *table, int16_t x);
 
 #endif
 //EOF
